Drop support for HDF5 1.6.
From the website: Release 1.8.5 is the latest release of HDF5 1.8. New projects
are strongly encouraged to use this release as it contains many new features,
file optimizations, and performance enhancements.

http://www.hdfgroup.org/HDF5/release/

The first new HDF5 1.8 features we will use is checking for the existence of
attributes and objects, and automatic creation of missing intermediate groups.