Drop helper functions using boost::bind.
These functions store a pointer on the data to access them at a later point in
program exectution. We do not support this practice any further as it
potentially causes a memory access violation.

Add helper functions for unique datasets which create/open a dataset on the fly
and write/read the data immediately.