test: reopen files properly by destroying the previous H5File object
Opening the file in read-only mode using the same H5File object after calling
close() without a preceding flush() results in a damaged file on disk.

The solution is to store the H5File object in a shared_ptr and to replace it by
a fresh instance for reading.

This issue was pointed out by Michael Kopp.