Fix maximally inefficient dataset chunk size.
For a dataset tracing the evolution of a single scalar (e.g. double), a chunk
size of 1×1 (8 bytes) would cause significant meta-data overhead, which would
lead to increased file size, and memory exhaustion upon subsequently reading
the dataset. As an example, reading an extensible dataset of 10⁶ double values
(7.6 MiB) with h5py would cause a segmentation fault due to low memory, on a
machine with 2 GiB memory.

Ensure that the chunk size is at least 8 KiB, by multiplying the outermost
dimension of chunk size by a power of 2. This algorithm is far from perfect,
since it ignores the other dimensions of the chunk, which would need to be
adjusted to allow e.g. tracing a single particle over time in a trajectory
of many particles, without reading the trajectory of all particles.

The minimum chunk size of 8 kiB is recommended in [1].

[1] http://www.hdfgroup.org/training/HDFtraining/UsersGuide/Perform.fm2.html

Note that HDF5 has an internal chunk cache, by default of size 1 MiB.
The compression filter is only applied when the chunk is written from
the cache to disk, which means that even in the case of appending data
smaller than the chunk size, the chunk will only be compressed and
written to disk once.

For details on chunking, see Chunking in HDF5 [2].

[2] http://www.hdfgroup.org/HDF5/doc/_topic/Chunking/