attribute: fix reading of string arrays
In the buffer filled by H5A_read(), the longest string is not
'\0'-terminated. Just converting the C strings to std::string resulted
in excessively long strings.

The issue was overlooked by the test since the longest string was the
last one (and the buffer was followed by a 0, obviously).