implemented filters and storage policies
The dataset layout policies contiguous, compact, and chunked
now support adding storage policies such as fill_value and
track_times via the "set" element function (ie fill_value is
not passed as a separate parameter to the dataset constructor
but rather together with a storage layout policy).
In addition, the chunked layout supports several filters
which can be activated using the "add" element function.

To store the policy class instances until they are activated
using the function "set_storage", std::vectors are used in
combination with boost::shared_ptrs (the original draft on how
the filters are to be handled did not work as we cannot store
derived class instances in vectors of base class type directly).

The constructor of the scaleoffset filter needs some work, I could
not get it to compile when using boost::enable_if<>.

Please see example/dataset.cpp for how usage may look like.